Extract types folder under src folder
