+++ src/types/d2.d.ts
